print_utils.c に print_* を追加
