+++ framework/exit_fatal.c
@@ -1,29 +1,5 @@
 #include "libunit.h"
 
-void	print_green(char *s, int fd)
-{
-	ft_putstr_fd(GREEN, fd);
-	ft_putstr_fd(s, fd);
-	ft_putstr_fd(DEFAULT, fd);
-	ft_putstr_fd("\n", fd);
-}
-
-void	print_red(char *s, int fd)
-{
-	ft_putstr_fd(RED, fd);
-	ft_putstr_fd(s, fd);
-	ft_putstr_fd(DEFAULT, fd);
-	ft_putstr_fd("\n", fd);
-}
-
-void	print_yellow(char *s, int fd)
-{
-	ft_putstr_fd(YELLOW, fd);
-	ft_putstr_fd(s, fd);
-	ft_putstr_fd(DEFAULT, fd);
-	ft_putstr_fd("\n", fd);
-}
-
 static int	ft_strlen(const char *s)
 {
 	int	i;
