+++ framework/libunit.h
@@ -34,6 +34,8 @@ void			exit_fatal(void);
 void			print_green(char *s, int fd);
 void			print_red(char *s, int fd);
 void			print_yellow(char *s, int fd);
+void			print_status_with_color(int exit_status, int fd);
+void			print_status(int exit_status, int fd);
 void			ft_putstr_fd(char *s, int fd);
 void			print_result(int checked, int num_tests, int fd);
